Add files via upload
In this program the user is asked to guess the word and if he/she guesses it in the given attempts then he/she will win or else he/she will lose the game. The list of words is already provided to the program and every time the program is run any random word will be chosen.